refactor: extract ui auth tenant helpers
